Gemini-python-ai-client/requirements.txt
# Install using: pip install -r requirements.txt


google-genai